api_server/requirements.txt
Django==3.2.7
django-redis==5.0.0
django-filter==2.4.0
django-markdown==0.8.4
django-model-utils==4.1.1
django-sslserver==0.22
djangorestframework==3.12.4
djangorestframework-api-key==2.0.0
django-oauth-toolkit==1.5.0
django-cors-headers==3.8.0
Pillow==8.3.2
psycopg2-binary==2.9.1
django-url-filter
